test(funput-engine): move public-API tests from lib.rs to tests/engine_api.rs
lib.rs carried ~345 lines of inline unit tests that only ever drive the public
Engine surface (Engine/Action/ImeResult) — no private items — so they are
integration tests in all but location. Move them into a new tests/engine_api.rs
and swap `use super::*` for `use funput_engine::{...}`.

This shrinks lib.rs from 589 to 243 lines (now just the frozen API + docs) and
puts the end-to-end Engine tests alongside the other themed integration groups.
The inline unit tests in the other src modules stay put on purpose: they cover
pub(crate)/private internals (diff, flip, process, should_restore, Session,
ImeResult::none/send) that integration tests cannot reach.

No test changed — same 115 total (31 unit + 84 integration, of which 28 are the
moved engine_api cases). cargo clippy --all-targets -D warnings and rustfmt clean.
